Reorganize project into package
