+++ vastvoid/__init__.py
@@ -1,0 +1,1 @@
+"""Core package for the VastVoid game."""
