+++ vastvoid/examples/black_plane.py
