+++ vastvoid/examples/colliding_star_systems.py
