+++ vastvoid/examples/nave.py
