+++ vastvoid/examples/space_objects.py
