+++ vastvoid/mechanics/combat.py
@@ -2,7 +2,7 @@ import math
 from dataclasses import dataclass, field
 from typing import List
 import pygame
-import config
+from .. import config
 
 
 @dataclass
