+++ vastvoid/mechanics/fraction.py
