+++ vastvoid/names.py
